backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML Dependencies
sentence-transformers==2.2.2
networkx==3.2.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
torch==2.1.1
deepdiff==6.7.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psutil==5.9.6

# Security
cryptography==41.0.8

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Background Tasks (for future use)
celery==5.3.4
redis==5.0.1